+++ example/main.dart
@@ -6,4 +6,6 @@ void main() {
 	print(carvable.apply()); // 'acd'
 
 	print('abcde'.carvable.remove(1, 2).remove(3, 4).apply()); // 'ace'
+
+  print('abd'.carvable.remove(2, 3).append('c').apply());
 }